# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ezamu Triad Mentorship Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users/onboard:
    post:
      operationId: onboardUser
      tags: [users]
      summary: Complete initial user onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardUserBody"
      responses:
        "200":
          description: Updated user after onboarding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /coaches:
    get:
      operationId: getCoaches
      tags: [coaches]
      summary: Get list of available coaches
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: field
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of coaches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coach"
  /coaches/{coachId}:
    get:
      operationId: getCoach
      tags: [coaches]
      summary: Get a specific coach profile
      parameters:
        - name: coachId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coach profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coach"
        "404":
          description: Coach not found
  /coaches/{coachId}/students:
    get:
      operationId: getCoachStudents
      tags: [coaches]
      summary: Get students assigned to a coach
      parameters:
        - name: coachId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages between current user and another
      parameters:
        - name: withUserId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by email (for chat/guardian invite)
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, coach, guardian]
        profilePicUrl:
          type: ["string", "null"]
        innerHeroArchetype:
          type: ["string", "null"]
        coachId:
          type: ["integer", "null"]
        peerId:
          type: ["integer", "null"]
        bio:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        fieldsOfInterest:
          type: array
          items:
            type: string
        fieldsOfExpertise:
          type: array
          items:
            type: string
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - firstName
        - lastName
        - email
        - role
        - fieldsOfInterest
        - fieldsOfExpertise
        - onboardingCompleted
        - createdAt
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
        profilePicUrl:
          type: ["string", "null"]
      required:
        - id
        - firstName
        - lastName
        - email
        - role
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        profilePicUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        fieldsOfInterest:
          type: array
          items:
            type: string
        fieldsOfExpertise:
          type: array
          items:
            type: string
    OnboardUserBody:
      type: object
      properties:
        age:
          type: integer
        bio:
          type: string
        fieldsOfInterest:
          type: array
          items:
            type: string
        fieldsOfExpertise:
          type: array
          items:
            type: string
      required:
        - bio
    Coach:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        profilePicUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        fieldsOfExpertise:
          type: array
          items:
            type: string
        studentCount:
          type: integer
      required:
        - id
        - firstName
        - lastName
        - email
        - fieldsOfExpertise
        - studentCount
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
        relatedId:
          type: ["integer", "null"]
      required:
        - id
        - userId
        - type
        - message
        - read
        - createdAt
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
        content:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
        senderName:
          type: string
        senderProfilePic:
          type: ["string", "null"]
      required:
        - id
        - senderId
        - receiverId
        - content
        - read
        - createdAt
        - senderName
    SendMessageBody:
      type: object
      properties:
        receiverId:
          type: integer
        content:
          type: string
      required:
        - receiverId
        - content
...