# lib/api-zod/src/generated/types/onboardUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ezamu Triad Mentorship Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardUserBody {
  age?: number;
  bio: string;
  fieldsOfInterest?: string[];
  fieldsOfExpertise?: string[];
}