# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ezamu Triad Mentorship Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  clerkId: string;
  firstName: string;
  lastName: string;
  email: string;
  role: UserRole;
  /** @nullable */
  profilePicUrl?: string | null;
  /** @nullable */
  innerHeroArchetype?: string | null;
  /** @nullable */
  coachId?: number | null;
  /** @nullable */
  peerId?: number | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  age?: number | null;
  fieldsOfInterest: string[];
  fieldsOfExpertise: string[];
  onboardingCompleted: boolean;
  createdAt: string;
}